requirements.txt
fastapi
uvicorn
mysql-connector-python
passlib[bcrypt]
pydantic-settings
pytest
httpx
